+++ shared-libs/ts-modules/start-core/lib/osBindings/DesiredStatus.ts
@@ -6,3 +6,4 @@ export type DesiredStatus =
   | { main: 'restarting'; restartAgain: boolean }
   | { main: 'running' }
   | { main: 'backing-up'; onComplete: StartStop }
+  | { main: 'updating'; onComplete: StartStop }